refactored tests and merged into one file
