+++ expr/func_aspercent_test.go
@@ -1,6 +1,7 @@
 package expr
 
 import (
+	"fmt"
 	"math"
 	"sort"
 	"strconv"
@@ -11,18 +12,328 @@ import (
 	"gopkg.in/raintank/schema.v1"
 )
 
+type errAsPercentNumSeriesMismatch struct {
+	numIn    int
+	numTotal int
+}
+
+func (e errAsPercentNumSeriesMismatch) Error() string {
+	return fmt.Sprintf("asPercent got %d input series but %d total series (should  be same amount or 1)", e.numIn, e.numTotal)
+}
+
+var a1 = []schema.Point{
+	{Val: 0, Ts: 10},
+	{Val: 50.6, Ts: 20},
+	{Val: 1234567890, Ts: 30},
+	{Val: math.NaN(), Ts: 40},
+}
+
+var b1 = []schema.Point{
+	{Val: 10, Ts: 10},
+	{Val: -25.8, Ts: 20},
+	{Val: float64(math.MaxFloat64 / 1000), Ts: 30},
+	{Val: math.NaN(), Ts: 40},
+}
+var a2 = []schema.Point{
+	{Val: 10, Ts: 10},
+	{Val: 25.3, Ts: 20},
+	{Val: 12345678900, Ts: 30},
+	{Val: -100, Ts: 40},
+}
+
+var b2 = []schema.Point{
+	{Val: -1000, Ts: 10},
+	{Val: 258, Ts: 20},
+	{Val: math.MaxFloat64, Ts: 30},
+	{Val: math.NaN(), Ts: 40},
+}
+
+var a1AsPercentOfa1 = []schema.Point{
+	{Val: math.NaN(), Ts: 10},
+	{Val: 100, Ts: 20},
+	{Val: 100, Ts: 30},
+	{Val: math.NaN(), Ts: 40},
+}
+var a1AsPercentOfa1b1 = []schema.Point{
+	{Val: 0, Ts: 10}, // 100 * 0 / (0+10)
+	{Val: float64(50.6) / float64(50.6-25.8) * 100, Ts: 20},
+	{Val: float64(1234567890) / (float64(1234567890) + math.MaxFloat64/1000) * 100, Ts: 30},
+	{Val: math.NaN(), Ts: 40},
+}
+var b1AsPercentOfa1b1 = []schema.Point{
+	{Val: 100, Ts: 10},
+	{Val: -100 * float64(25.8) / float64(50.6-25.8), Ts: 20},
+	{Val: 100 * (math.MaxFloat64 / 1000) / (1234567890 + math.MaxFloat64/1000), Ts: 30},
+	{Val: math.NaN(), Ts: 40},
+}
+var b1AsPercentOfa1 = []schema.Point{
+	{Val: math.NaN(), Ts: 10},
+	{Val: float64(-25.8) / 50.6 * 100, Ts: 20},
+	{Val: float64(math.MaxFloat64/1000.0) / 1234567890 * 100, Ts: 30},
+	{Val: math.NaN(), Ts: 40},
+}
+
+var a1AsPercentOfa2 = []schema.Point{
+	{Val: 0, Ts: 10},
+	{Val: 200, Ts: 20},
+	{Val: 10, Ts: 30},
+	{Val: math.NaN(), Ts: 40},
+}
+var b1AsPercentOfb2 = []schema.Point{
+	{Val: -1, Ts: 10},
+	{Val: -10, Ts: 20},
+	{Val: 0.1, Ts: 30},
+	{Val: math.NaN(), Ts: 40},
+}
+
+var a1AsPercentOfFloat10 = []schema.Point{
+	{Val: 0, Ts: 10},
+	{Val: float64(50.6) / 10.0 * 100, Ts: 20},
+	{Val: 12345678900, Ts: 30},
+	{Val: math.NaN(), Ts: 40},
+}
+
+// first, the case of no value for "total" specified, which should use the sum of all input series as total series
+// let's try 3 different ways of inputting series
+
+func TestAsPercentSingleInputUsingSelf(t *testing.T) {
+	f := getNewAsPercent(
+		[]models.Series{
+			{
+				Interval:   10,
+				QueryPatt:  "a",
+				Target:     "a",
+				Datapoints: getCopy(a1),
+			},
+		})
+	out := []models.Series{
+		{
+			Interval:   10,
+			QueryPatt:  "asPercent(a,sumSeries(a))",
+			Target:     "asPercent(a,sumSeries(a))",
+			Datapoints: getCopy(a1AsPercentOfa1),
+		},
+	}
+	got, err := f.Exec(make(map[Req][]models.Series))
+	if err := equalOutput(out, got, nil, err); err != nil {
+		t.Fatal(err)
+	}
+}
+
+func TestAsPercentDoubleInputUsingSelf(t *testing.T) {
+	f := getNewAsPercent(
+		[]models.Series{
+			{
+				Interval:   10,
+				QueryPatt:  "a.*",
+				Target:     "a.a",
+				Datapoints: getCopy(a1),
+			},
+			{
+				Interval:   10,
+				QueryPatt:  "a.*",
+				Target:     "a.b",
+				Datapoints: getCopy(b1),
+			},
+		})
+	out := []models.Series{
+		{
+			Interval:   10,
+			QueryPatt:  "asPercent(a.*,sumSeries(a.*))",
+			Target:     "asPercent(a.a,sumSeries(a.*))",
+			Datapoints: getCopy(a1AsPercentOfa1b1),
+		},
+		{
+			Interval:   10,
+			QueryPatt:  "asPercent(a.*,sumSeries(a.*))",
+			Target:     "asPercent(a.b,sumSeries(a.*))",
+			Datapoints: getCopy(b1AsPercentOfa1b1),
+		},
+	}
+	got, err := f.Exec(make(map[Req][]models.Series))
+	if err := equalOutput(out, got, nil, err); err != nil {
+		t.Fatal(err)
+	}
+}
+
+// now let's try the case of specifying another seriesList.
+// specifically, 1 series, same amount of series as inputs (2), and different amount (3)
+// for simplicity, each case uses 2 inputs.
+
+func TestAsPercentDoubleInputUsingOne(t *testing.T) {
+	f := getNewAsPercent(
+		[]models.Series{
+			{
+				Interval:   10,
+				QueryPatt:  "a.*",
+				Target:     "a.a",
+				Datapoints: getCopy(a1),
+			},
+			{
+				Interval:   10,
+				QueryPatt:  "a.*",
+				Target:     "a.b",
+				Datapoints: getCopy(b1),
+			},
+		})
+	f.totalSeries = NewMock([]models.Series{{
+		Interval:   10,
+		QueryPatt:  "foo.*",
+		Target:     "foo.a",
+		Datapoints: getCopy(a1),
+	}})
+	out := []models.Series{
+		{
+			Interval:   10,
+			QueryPatt:  "asPercent(a.*,foo.*)",
+			Target:     "asPercent(a.a,foo.a)",
+			Datapoints: getCopy(a1AsPercentOfa1),
+		},
+		{
+			Interval:   10,
+			QueryPatt:  "asPercent(a.*,foo.*)",
+			Target:     "asPercent(a.b,foo.a)",
+			Datapoints: getCopy(b1AsPercentOfa1),
+		},
+	}
+	got, err := f.Exec(make(map[Req][]models.Series))
+	if err := equalOutput(out, got, nil, err); err != nil {
+		t.Fatal(err)
+	}
+}
+func TestAsPercentDoubleInputUsingTwo(t *testing.T) {
+	f := getNewAsPercent(
+		[]models.Series{
+			{
+				Interval:   10,
+				QueryPatt:  "a.*",
+				Target:     "a.a",
+				Datapoints: getCopy(a1),
+			},
+			{
+				Interval:   10,
+				QueryPatt:  "a.*",
+				Target:     "a.b",
+				Datapoints: getCopy(b1),
+			},
+		})
+	f.totalSeries = NewMock([]models.Series{
+		{
+			Interval:   10,
+			QueryPatt:  "total.*",
+			Target:     "total.a",
+			Datapoints: getCopy(a2),
+		},
+		{
+			Interval:   10,
+			QueryPatt:  "total.*",
+			Target:     "total.b",
+			Datapoints: getCopy(b2),
+		},
+	})
+	out := []models.Series{
+		{
+			Interval:   10,
+			QueryPatt:  "asPercent(a.*,total.*)",
+			Target:     "asPercent(a.a,total.a)",
+			Datapoints: getCopy(a1AsPercentOfa2),
+		},
+		{
+			Interval:   10,
+			QueryPatt:  "asPercent(a.*,total.*)",
+			Target:     "asPercent(a.b,total.b)",
+			Datapoints: getCopy(b1AsPercentOfb2),
+		},
+	}
+	got, err := f.Exec(make(map[Req][]models.Series))
+	if err := equalOutput(out, got, nil, err); err != nil {
+		t.Fatal(err)
+	}
+}
+
+// 2 input series, 3 total series -> not allowed!
+func TestAsPercentDoubleInputUsingThree(t *testing.T) {
+	f := getNewAsPercent(
+		[]models.Series{
+			{
+				Interval:   10,
+				QueryPatt:  "a.*",
+				Target:     "a.a",
+				Datapoints: getCopy(a1),
+			},
+			{
+				Interval:   10,
+				QueryPatt:  "a.*",
+				Target:     "a.b",
+				Datapoints: getCopy(b1),
+			},
+		})
+	f.totalSeries = NewMock([]models.Series{
+		{
+			Interval:   10,
+			QueryPatt:  "total.*",
+			Target:     "total.a",
+			Datapoints: getCopy(a2),
+		},
+		{
+			Interval:   10,
+			QueryPatt:  "total.*",
+			Target:     "total.b",
+			Datapoints: getCopy(b2),
+		},
+		{
+			Interval:   10,
+			QueryPatt:  "total.*",
+			Target:     "total.c",
+			Datapoints: getCopy(b2),
+		},
+	})
+	out := []models.Series{}
+	got, err := f.Exec(make(map[Req][]models.Series))
+	expErr := errAsPercentNumSeriesMismatch{2, 3}
+	if err := equalOutput(out, got, expErr, err); err != nil {
+		t.Fatal(err)
+	}
+}
+
+// finally, test the case where an integer is specified as the total
+// for simplicity, we'll just use 1 input.
+
+func TestAsPercentSingleInputUsingFloat10(t *testing.T) {
+	f := getNewAsPercent(
+		[]models.Series{
+			{
+				Interval:   10,
+				QueryPatt:  "a",
+				Target:     "a",
+				Datapoints: getCopy(a1),
+			},
+		})
+	f.totalFloat = 10
+	out := []models.Series{
+		{
+			Interval:   10,
+			QueryPatt:  "asPercent(a,10)",
+			Target:     "asPercent(a,10)",
+			Datapoints: getCopy(a1AsPercentOfFloat10),
+		},
+	}
+	got, err := f.Exec(make(map[Req][]models.Series))
+	if err := equalOutput(out, got, nil, err); err != nil {
+		t.Fatal(err)
+	}
+}
+
+func getNewAsPercent(in []models.Series) *FuncAsPercent {
+	f := NewAsPercent()
+	ps := f.(*FuncAsPercent)
+	ps.in = NewMock(in)
+	return ps
+}
+
 func TestAsPercentSingleNoArg(t *testing.T) {
 
-	out := []schema.Point{
-		{Val: math.NaN(), Ts: 10},
-		{Val: math.NaN(), Ts: 20},
-		{Val: 100, Ts: 30},
-		{Val: math.NaN(), Ts: 40},
-		{Val: math.NaN(), Ts: 50},
-		{Val: 100, Ts: 60},
-	}
-	testAsPercent(
-		"single-noarg",
+	f := getNewAsPercent(
 		[]models.Series{
 			{
 				Interval:   10,
@@ -31,20 +342,28 @@ func TestAsPercentSingleNoArg(t *testing.T) {
 				Datapoints: getCopy(a),
 			},
 		},
-		[]models.Series{
-			{
-				Interval:   10,
-				QueryPatt:  "asPercent(func(tag=something;tag2=anything),sumSeries(func(tag=something;tag2=anything)))",
-				Target:     "asPercent(a;tag=something;tag2=anything,sumSeries(func(tag=something;tag2=anything)))",
-				Datapoints: out,
+	)
+	out := []models.Series{
+		{
+			Interval:  10,
+			QueryPatt: "asPercent(func(tag=something;tag2=anything),sumSeries(func(tag=something;tag2=anything)))",
+			Target:    "asPercent(a;tag=something;tag2=anything,sumSeries(func(tag=something;tag2=anything)))",
+			Datapoints: []schema.Point{
+				{Val: math.NaN(), Ts: 10},
+				{Val: math.NaN(), Ts: 20},
+				{Val: 100, Ts: 30},
+				{Val: math.NaN(), Ts: 40},
+				{Val: math.NaN(), Ts: 50},
+				{Val: 100, Ts: 60},
 			},
 		},
-		t,
-		math.NaN(),
-		nil,
-		nil,
-		"None",
-	)
+	}
+
+	got, err := f.Exec(make(map[Req][]models.Series))
+	if err := equalOutput(out, got, nil, err); err != nil {
+		t.Fatal(err)
+	}
+
 }
 
 func TestAsPercentMultipleNoArg(t *testing.T) {
@@ -64,8 +383,8 @@ func TestAsPercentMultipleNoArg(t *testing.T) {
 		{Val: 100, Ts: 50},
 		{Val: math.NaN(), Ts: 60},
 	}
-	testAsPercent(
-		"multi-noarg",
+
+	f := getNewAsPercent(
 		[]models.Series{
 			{
 				Interval:   10,
@@ -80,40 +399,31 @@ func TestAsPercentMultipleNoArg(t *testing.T) {
 				Datapoints: getCopy(b),
 			},
 		},
-		[]models.Series{
-			{
-				Interval:   10,
-				QueryPatt:  "asPercent(func(tag=something;tag2=anything),sumSeries(func(tag=something;tag2=anything)))",
-				Target:     "asPercent(a;tag=something;tag2=anything,sumSeries(func(tag=something;tag2=anything)))",
-				Datapoints: out1,
-			},
-			{
-				Interval:   10,
-				QueryPatt:  "asPercent(func(tag=something;tag2=anything),sumSeries(func(tag=something;tag2=anything)))",
-				Target:     "asPercent(b;tag=something;tag2=anything,sumSeries(func(tag=something;tag2=anything)))",
-				Datapoints: out2,
-			},
-		},
-		t,
-		math.NaN(),
-		nil,
-		nil,
-		"None",
 	)
+	out := []models.Series{
+		{
+			Interval:   10,
+			QueryPatt:  "asPercent(func(tag=something;tag2=anything),sumSeries(func(tag=something;tag2=anything)))",
+			Target:     "asPercent(a;tag=something;tag2=anything,sumSeries(func(tag=something;tag2=anything)))",
+			Datapoints: out1,
+		},
+		{
+			Interval:   10,
+			QueryPatt:  "asPercent(func(tag=something;tag2=anything),sumSeries(func(tag=something;tag2=anything)))",
+			Target:     "asPercent(b;tag=something;tag2=anything,sumSeries(func(tag=something;tag2=anything)))",
+			Datapoints: out2,
+		},
+	}
+
+	got, err := f.Exec(make(map[Req][]models.Series))
+	if err := equalOutput(out, got, nil, err); err != nil {
+		t.Fatal(err)
+	}
+
 }
 
 func TestAsPercentTotalFloat(t *testing.T) {
-	total := 123.456
-	out := []schema.Point{
-		{Val: 0, Ts: 10},
-		{Val: 0, Ts: 20},
-		{Val: 5.5 / total * 100, Ts: 30},
-		{Val: math.NaN(), Ts: 40},
-		{Val: math.NaN(), Ts: 50},
-		{Val: 1234567890 / total * 100, Ts: 60},
-	}
-	testAsPercent(
-		"single-noarg",
+	f := getNewAsPercent(
 		[]models.Series{
 			{
 				Interval:   10,
@@ -122,20 +432,30 @@ func TestAsPercentTotalFloat(t *testing.T) {
 				Datapoints: getCopy(a),
 			},
 		},
-		[]models.Series{
-			{
-				Interval:   10,
-				QueryPatt:  "asPercent(func(a;tag=something;tag2=anything),123.456)",
-				Target:     "asPercent(a;tag=something;tag2=anything,123.456)",
-				Datapoints: out,
+	)
+
+	f.totalFloat = 123.456
+	out := []models.Series{
+		{
+			Interval:  10,
+			QueryPatt: "asPercent(func(a;tag=something;tag2=anything),123.456)",
+			Target:    "asPercent(a;tag=something;tag2=anything,123.456)",
+			Datapoints: []schema.Point{
+				{Val: 0, Ts: 10},
+				{Val: 0, Ts: 20},
+				{Val: 5.5 / f.totalFloat * 100, Ts: 30},
+				{Val: math.NaN(), Ts: 40},
+				{Val: math.NaN(), Ts: 50},
+				{Val: 1234567890 / f.totalFloat * 100, Ts: 60},
 			},
 		},
-		t,
-		total,
-		nil,
-		nil,
-		"123.456",
-	)
+	}
+
+	got, err := f.Exec(make(map[Req][]models.Series))
+	if err := equalOutput(out, got, nil, err); err != nil {
+		t.Fatal(err)
+	}
+
 }
 
 func TestAsPercentTotalSerie(t *testing.T) {
@@ -155,8 +475,7 @@ func TestAsPercentTotalSerie(t *testing.T) {
 		{Val: math.NaN(), Ts: 50},
 		{Val: float64(250) / 1234567890 * 100, Ts: 60},
 	}
-	testAsPercent(
-		"multi-serietotal",
+	f := getNewAsPercent(
 		[]models.Series{
 			{
 				Interval:   10,
@@ -171,22 +490,8 @@ func TestAsPercentTotalSerie(t *testing.T) {
 				Datapoints: getCopy(d),
 			},
 		},
-		[]models.Series{
-			{
-				Interval:   10,
-				QueryPatt:  "asPercent(func(tag=something;tag2=anything),func(tag=some;tag2=totalSerie))",
-				Target:     "asPercent(b;tag=something;tag2=anything,a;tag=some;tag2=totalSerie)",
-				Datapoints: out1,
-			},
-			{
-				Interval:   10,
-				QueryPatt:  "asPercent(func(tag=something;tag2=anything),func(tag=some;tag2=totalSerie))",
-				Target:     "asPercent(d;tag=something;tag2=anything,a;tag=some;tag2=totalSerie)",
-				Datapoints: out2,
-			},
-		},
-		t,
-		math.NaN(),
+	)
+	f.totalSeries = NewMock(
 		[]models.Series{
 			{
 				Interval:   10,
@@ -195,9 +500,26 @@ func TestAsPercentTotalSerie(t *testing.T) {
 				Datapoints: getCopy(a),
 			},
 		},
-		nil,
-		"serie",
 	)
+	out := []models.Series{
+		{
+			Interval:   10,
+			QueryPatt:  "asPercent(func(tag=something;tag2=anything),func(tag=some;tag2=totalSerie))",
+			Target:     "asPercent(b;tag=something;tag2=anything,a;tag=some;tag2=totalSerie)",
+			Datapoints: out1,
+		},
+		{
+			Interval:   10,
+			QueryPatt:  "asPercent(func(tag=something;tag2=anything),func(tag=some;tag2=totalSerie))",
+			Target:     "asPercent(d;tag=something;tag2=anything,a;tag=some;tag2=totalSerie)",
+			Datapoints: out2,
+		},
+	}
+
+	got, err := f.Exec(make(map[Req][]models.Series))
+	if err := equalOutput(out, got, nil, err); err != nil {
+		t.Fatal(err)
+	}
 }
 
 func TestAsPercentTotalSeries(t *testing.T) {
@@ -217,8 +539,8 @@ func TestAsPercentTotalSeries(t *testing.T) {
 		{Val: float64(80) / 3.0 * 100, Ts: 50},
 		{Val: float64(250) / 4 * 100, Ts: 60},
 	}
-	testAsPercent(
-		"multi-seriestotal",
+
+	f := getNewAsPercent(
 		[]models.Series{
 			{
 				Interval:   10,
@@ -233,22 +555,8 @@ func TestAsPercentTotalSeries(t *testing.T) {
 				Datapoints: getCopy(d),
 			},
 		},
-		[]models.Series{
-			{
-				Interval:   10,
-				QueryPatt:  "asPercent(func(tag=something;tag2=anything),func(tag=some;tag2=totalSerie))",
-				Target:     "asPercent(b;tag=something;tag2=anything,a;tag=some;tag2=totalSerie)",
-				Datapoints: out1,
-			},
-			{
-				Interval:   10,
-				QueryPatt:  "asPercent(func(tag=something;tag2=anything),func(tag=some;tag2=totalSerie))",
-				Target:     "asPercent(d;tag=something;tag2=anything,c;tag=some;tag2=totalSerie)",
-				Datapoints: out2,
-			},
-		},
-		t,
-		math.NaN(),
+	)
+	f.totalSeries = NewMock(
 		[]models.Series{
 			{
 				Interval:   10,
@@ -263,9 +571,26 @@ func TestAsPercentTotalSeries(t *testing.T) {
 				Datapoints: getCopy(a),
 			},
 		},
-		nil,
-		"series",
 	)
+	out := []models.Series{
+		{
+			Interval:   10,
+			QueryPatt:  "asPercent(func(tag=something;tag2=anything),func(tag=some;tag2=totalSerie))",
+			Target:     "asPercent(b;tag=something;tag2=anything,a;tag=some;tag2=totalSerie)",
+			Datapoints: out1,
+		},
+		{
+			Interval:   10,
+			QueryPatt:  "asPercent(func(tag=something;tag2=anything),func(tag=some;tag2=totalSerie))",
+			Target:     "asPercent(d;tag=something;tag2=anything,c;tag=some;tag2=totalSerie)",
+			Datapoints: out2,
+		},
+	}
+
+	got, err := f.Exec(make(map[Req][]models.Series))
+	if err := equalOutput(out, got, nil, err); err != nil {
+		t.Fatal(err)
+	}
 }
 
 func TestAsPercentNoArgNodes(t *testing.T) {
@@ -293,8 +618,8 @@ func TestAsPercentNoArgNodes(t *testing.T) {
 		{Val: 100, Ts: 50},
 		{Val: 100, Ts: 60},
 	}
-	testAsPercent(
-		"multi-seriebynode",
+
+	f := getNewAsPercent(
 		[]models.Series{
 			{
 				Interval:   10,
@@ -315,32 +640,36 @@ func TestAsPercentNoArgNodes(t *testing.T) {
 				Datapoints: getCopy(c),
 			},
 		},
-		[]models.Series{
-			{
-				Interval:   10,
-				QueryPatt:  "asPercent(func(tag=something;tag2=anything),sumSeries(func(tag=something;tag2=anything)))",
-				Target:     "asPercent(this.that.a;tag=something;tag2=anything,sumSeries(func(tag=something;tag2=anything)))",
-				Datapoints: out1,
-			},
-			{
-				Interval:   10,
-				QueryPatt:  "asPercent(func(tag=something;tag2=anything),sumSeries(func(tag=something;tag2=anything)))",
-				Target:     "asPercent(this.that.b;tag=something;tag2=anything,sumSeries(func(tag=something;tag2=anything)))",
-				Datapoints: out2,
-			},
-			{
-				Interval:   10,
-				QueryPatt:  "asPercent(func(tag=something;tag2=anything),func(tag=something;tag2=anything))",
-				Target:     "asPercent(this.this.c;tag=something;tag2=anything,this.this.c;tag=something;tag2=anything)",
-				Datapoints: out3,
-			},
-		},
-		t,
-		math.NaN(),
-		nil,
-		[]expr{{etype: etFloat, float: 0}, {etype: etInt, int: 1}},
-		"None",
 	)
+	f.nodes = []expr{{etype: etFloat, float: 0}, {etype: etInt, int: 1}}
+	out := []models.Series{
+		{
+			Interval:   10,
+			QueryPatt:  "asPercent(func(tag=something;tag2=anything),sumSeries(func(tag=something;tag2=anything)))",
+			Target:     "asPercent(this.that.a;tag=something;tag2=anything,sumSeries(func(tag=something;tag2=anything)))",
+			Datapoints: out1,
+		},
+		{
+			Interval:   10,
+			QueryPatt:  "asPercent(func(tag=something;tag2=anything),sumSeries(func(tag=something;tag2=anything)))",
+			Target:     "asPercent(this.that.b;tag=something;tag2=anything,sumSeries(func(tag=something;tag2=anything)))",
+			Datapoints: out2,
+		},
+		{
+			Interval:   10,
+			QueryPatt:  "asPercent(func(tag=something;tag2=anything),func(tag=something;tag2=anything))",
+			Target:     "asPercent(this.this.c;tag=something;tag2=anything,this.this.c;tag=something;tag2=anything)",
+			Datapoints: out3,
+		},
+	}
+
+	got, err := f.Exec(make(map[Req][]models.Series))
+	sort.Slice(got, func(i, j int) bool { return got[i].Target < got[j].Target })
+	sort.Slice(out, func(i, j int) bool { return out[i].Target < out[j].Target })
+	if err := equalOutput(out, got, nil, err); err != nil {
+		t.Fatal(err)
+	}
+
 }
 
 func TestAsPercentNoArgTagNodes(t *testing.T) {
@@ -368,8 +697,8 @@ func TestAsPercentNoArgTagNodes(t *testing.T) {
 		{Val: 100, Ts: 50},
 		{Val: 100, Ts: 60},
 	}
-	testAsPercent(
-		"multi-seriebynode",
+
+	f := getNewAsPercent(
 		[]models.Series{
 			{
 				Interval:   10,
@@ -390,32 +719,36 @@ func TestAsPercentNoArgTagNodes(t *testing.T) {
 				Datapoints: getCopy(c),
 			},
 		},
-		[]models.Series{
-			{
-				Interval:   10,
-				QueryPatt:  "asPercent(func(tag=something1;tag2=anything),sumSeries(func(tag=something1;tag2=anything)))",
-				Target:     "asPercent(this.that.a;tag=something1;tag2=anything,sumSeries(func(tag=something1;tag2=anything)))",
-				Datapoints: out1,
-			},
-			{
-				Interval:   10,
-				QueryPatt:  "asPercent(func(tag=something1;tag2=anything),sumSeries(func(tag=something1;tag2=anything)))",
-				Target:     "asPercent(this.those.b;tag=something1;tag2=anything,sumSeries(func(tag=something1;tag2=anything)))",
-				Datapoints: out2,
-			},
-			{
-				Interval:   10,
-				QueryPatt:  "asPercent(func(tag=something2;tag2=anything),func(tag=something2;tag2=anything))",
-				Target:     "asPercent(this.this.c;tag=something2;tag2=anything,this.this.c;tag=something2;tag2=anything)",
-				Datapoints: out3,
-			},
-		},
-		t,
-		math.NaN(),
-		nil,
-		[]expr{{etype: etString, str: "tag"}},
-		"None",
 	)
+	f.nodes = []expr{{etype: etString, str: "tag"}}
+	out := []models.Series{
+		{
+			Interval:   10,
+			QueryPatt:  "asPercent(func(tag=something1;tag2=anything),sumSeries(func(tag=something1;tag2=anything)))",
+			Target:     "asPercent(this.that.a;tag=something1;tag2=anything,sumSeries(func(tag=something1;tag2=anything)))",
+			Datapoints: out1,
+		},
+		{
+			Interval:   10,
+			QueryPatt:  "asPercent(func(tag=something1;tag2=anything),sumSeries(func(tag=something1;tag2=anything)))",
+			Target:     "asPercent(this.those.b;tag=something1;tag2=anything,sumSeries(func(tag=something1;tag2=anything)))",
+			Datapoints: out2,
+		},
+		{
+			Interval:   10,
+			QueryPatt:  "asPercent(func(tag=something2;tag2=anything),func(tag=something2;tag2=anything))",
+			Target:     "asPercent(this.this.c;tag=something2;tag2=anything,this.this.c;tag=something2;tag2=anything)",
+			Datapoints: out3,
+		},
+	}
+
+	got, err := f.Exec(make(map[Req][]models.Series))
+	sort.Slice(got, func(i, j int) bool { return got[i].Target < got[j].Target })
+	sort.Slice(out, func(i, j int) bool { return out[i].Target < out[j].Target })
+	if err := equalOutput(out, got, nil, err); err != nil {
+		t.Fatal(err)
+	}
+
 }
 
 func TestAsPercentSeriesByNodes(t *testing.T) {
@@ -443,8 +776,7 @@ func TestAsPercentSeriesByNodes(t *testing.T) {
 		{Val: math.NaN(), Ts: 50},
 		{Val: math.NaN(), Ts: 60},
 	}
-	testAsPercent(
-		"multi-seriebynodeandseries",
+	f := getNewAsPercent(
 		[]models.Series{
 			{
 				Interval:   10,
@@ -465,34 +797,8 @@ func TestAsPercentSeriesByNodes(t *testing.T) {
 				Datapoints: getCopy(c),
 			},
 		},
-		[]models.Series{
-			{
-				Interval:   10,
-				QueryPatt:  "asPercent(func(tag=something;tag2=anything),sumSeries(func(tag=something;tag2=totalSerie)))",
-				Target:     "asPercent(this.that.a;tag=something;tag2=anything,sumSeries(func(tag=something;tag2=totalSerie)))",
-				Datapoints: out1,
-			},
-			{
-				Interval:   10,
-				QueryPatt:  "asPercent(func(tag=something;tag2=anything),sumSeries(func(tag=something;tag2=totalSerie)))",
-				Target:     "asPercent(this.that.b;tag=something;tag2=anything,sumSeries(func(tag=something;tag2=totalSerie)))",
-				Datapoints: out2,
-			},
-			{
-				Interval:   10,
-				QueryPatt:  "asPercent(func(tag=something;tag2=anything),MISSING)",
-				Target:     "asPercent(this.this.c;tag=something;tag2=anything,MISSING)",
-				Datapoints: allNaN,
-			},
-			{
-				Interval:   10,
-				QueryPatt:  "asPercent(MISSING,func(tag=something;tag2=totalSerie))",
-				Target:     "asPercent(MISSING,this.those.ab;tag=something;tag2=totalSerie)",
-				Datapoints: allNaN,
-			},
-		},
-		t,
-		math.NaN(),
+	)
+	f.totalSeries = NewMock(
 		[]models.Series{
 			{
 				Interval:   10,
@@ -513,75 +819,42 @@ func TestAsPercentSeriesByNodes(t *testing.T) {
 				Datapoints: getCopy(sumcd),
 			},
 		},
-		[]expr{{etype: etFloat, float: 0}, {etype: etInt, int: 1}},
-		"series",
 	)
-}
+	f.nodes = []expr{{etype: etFloat, float: 0}, {etype: etInt, int: 1}}
+	out := []models.Series{
+		{
+			Interval:   10,
+			QueryPatt:  "asPercent(func(tag=something;tag2=anything),sumSeries(func(tag=something;tag2=totalSerie)))",
+			Target:     "asPercent(this.that.a;tag=something;tag2=anything,sumSeries(func(tag=something;tag2=totalSerie)))",
+			Datapoints: out1,
+		},
+		{
+			Interval:   10,
+			QueryPatt:  "asPercent(func(tag=something;tag2=anything),sumSeries(func(tag=something;tag2=totalSerie)))",
+			Target:     "asPercent(this.that.b;tag=something;tag2=anything,sumSeries(func(tag=something;tag2=totalSerie)))",
+			Datapoints: out2,
+		},
+		{
+			Interval:   10,
+			QueryPatt:  "asPercent(func(tag=something;tag2=anything),MISSING)",
+			Target:     "asPercent(this.this.c;tag=something;tag2=anything,MISSING)",
+			Datapoints: allNaN,
+		},
+		{
+			Interval:   10,
+			QueryPatt:  "asPercent(MISSING,func(tag=something;tag2=totalSerie))",
+			Target:     "asPercent(MISSING,this.those.ab;tag=something;tag2=totalSerie)",
+			Datapoints: allNaN,
+		},
+	}
 
-func testAsPercent(name string, in []models.Series, out []models.Series, t *testing.T, totalFloat float64, totalSeries []models.Series, nodes []expr, total string) {
-	f := NewAsPercent()
-	f.(*FuncAsPercent).in = NewMock(in)
-	if totalSeries != nil {
-		f.(*FuncAsPercent).totalSeries = NewMock(totalSeries)
-	}
-	f.(*FuncAsPercent).totalFloat = totalFloat
-	f.(*FuncAsPercent).nodes = nodes
-
-	originalSeries := make([]models.Series, len(in))
-	for i, serie := range in {
-		originalSeries[i] = serie.Copy(make([]schema.Point, 0, len(serie.Datapoints)))
-	}
-	gots, err := f.Exec(make(map[Req][]models.Series))
-	if err != nil {
-		t.Fatalf("case %q (%q, %v): err should be nil. got %q", name, total, nodes, err)
-	}
-	sort.Slice(gots, func(i, j int) bool { return gots[i].Target < gots[j].Target })
+	got, err := f.Exec(make(map[Req][]models.Series))
+	sort.Slice(got, func(i, j int) bool { return got[i].Target < got[j].Target })
 	sort.Slice(out, func(i, j int) bool { return out[i].Target < out[j].Target })
-	if len(gots) != len(out) {
-		t.Fatalf("case %q (%q, %v): asPercent len output expected %d, got %d", name, total, nodes, len(out), len(gots))
-	}
-	for i, g := range gots {
-		exp := out[i]
-		if g.Target != exp.Target {
-			t.Fatalf("case %q (%q, %v): expected target %q, got %q", name, total, nodes, exp.Target, g.Target)
-		}
-		if g.QueryPatt != exp.QueryPatt {
-			t.Fatalf("case %q (%q, %v),: expected querypatt %q, got %q", name, total, nodes, exp.QueryPatt, g.QueryPatt)
-		}
-		if len(g.Datapoints) != len(exp.Datapoints) {
-			t.Fatalf("case %q (%q, %v): len output expected %d, got %d", name, total, nodes, len(exp.Datapoints), len(g.Datapoints))
-		}
-		for j, p := range exp.Datapoints {
-			bothNaN := math.IsNaN(p.Val) && math.IsNaN(g.Datapoints[j].Val)
-			if (bothNaN || p.Val == g.Datapoints[j].Val) && p.Ts == g.Datapoints[j].Ts {
-				continue
-			}
-
-			t.Fatalf("case %q (%q, %v): output point %d - expected %v got %v", name, total, nodes, j, p, g.Datapoints[j])
-		}
+	if err := equalOutput(out, got, nil, err); err != nil {
+		t.Fatal(err)
 	}
 
-	// Test if original series was modified
-	for i, orig := range originalSeries {
-		inSerie := in[i]
-		if orig.Target != inSerie.Target {
-			t.Fatalf("case %q (%q, %v): put in target %q, ended with %q", name, total, nodes, orig.Target, inSerie.Target)
-		}
-		if orig.QueryPatt != inSerie.QueryPatt {
-			t.Fatalf("case %q (%q, %v),: put in querypatt %q, ended with %q", name, total, nodes, orig.Target, inSerie.Target)
-		}
-		if len(orig.Datapoints) != len(inSerie.Datapoints) {
-			t.Fatalf("case %q (%q, %v): put in len %d, ended with %d", name, total, nodes, len(inSerie.Datapoints), len(orig.Datapoints))
-		}
-		for j, p := range inSerie.Datapoints {
-			bothNaN := math.IsNaN(p.Val) && math.IsNaN(orig.Datapoints[j].Val)
-			if (bothNaN || p.Val == orig.Datapoints[j].Val) && p.Ts == orig.Datapoints[j].Ts {
-				continue
-			}
-
-			t.Fatalf("case %q (%q, %v): output point %d - put in %v ended with %v", name, total, nodes, j, p, orig.Datapoints[j])
-		}
-	}
 }
 
 func BenchmarkAsPercent10k_1NoNulls(b *testing.B) {
